pyproject.toml
[project]
name = "insightfeed"
version = "0.1.0"
description = ""
authors = [
    {name = "Berat Polat",email = "119815205+BrplT0@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "feedparser (>=6.0.12,<7.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "pgvector (>=0.4.2,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
